Move conversation to it's own module
