+++ src/codex_agent.rs
@@ -3,22 +3,18 @@ use agent_client_protocol::{
     AuthenticateResponse, BlobResourceContents, CancelNotification, Client, ContentBlock, Diff,
     EmbeddedResource, EmbeddedResourceResource, Error, ImageContent, InitializeRequest,
     InitializeResponse, LoadSessionRequest, LoadSessionResponse, McpCapabilities, McpServer,
-    ModelId, ModelInfo, NewSessionRequest, NewSessionResponse, PermissionOption,
-    PermissionOptionId, PermissionOptionKind, Plan, PlanEntry, PlanEntryPriority, PlanEntryStatus,
-    PromptCapabilities, PromptRequest, PromptResponse, ReadTextFileRequest,
-    RequestPermissionOutcome, RequestPermissionRequest, ResourceLink, SessionId, SessionMode,
-    SessionModeId, SessionModeState, SessionModelState, SessionNotification, SessionUpdate,
+    NewSessionRequest, NewSessionResponse, PermissionOption, PermissionOptionId,
+    PermissionOptionKind, Plan, PlanEntry, PlanEntryPriority, PlanEntryStatus, PromptCapabilities,
+    PromptRequest, PromptResponse, ReadTextFileRequest, RequestPermissionOutcome,
+    RequestPermissionRequest, ResourceLink, SessionId, SessionNotification, SessionUpdate,
     SetSessionModeRequest, SetSessionModeResponse, SetSessionModelRequest, SetSessionModelResponse,
     StopReason, TerminalId, TextContent, TextResourceContents, ToolCall, ToolCallContent,
     ToolCallId, ToolCallLocation, ToolCallStatus, ToolCallUpdate, ToolCallUpdateFields, ToolKind,
     V1, WriteTextFileRequest,
 };
-use codex_common::{
-    approval_presets::{ApprovalPreset, builtin_approval_presets},
-    model_presets::{ModelPreset, builtin_model_presets},
-};
+use codex_common::model_presets::{ModelPreset, builtin_model_presets};
 use codex_core::{
-    CodexConversation, ConversationManager, NewConversation,
+    ConversationManager, NewConversation,
     auth::{AuthManager, CodexAuth, login_with_api_key},
     config::Config,
     config_types::{McpServerConfig, McpServerTransportConfig},
@@ -26,18 +22,14 @@ use codex_core::{
     protocol::{
         AgentMessageDeltaEvent, AgentMessageEvent, AgentReasoningDeltaEvent, AgentReasoningEvent,
         AgentReasoningRawContentDeltaEvent, AgentReasoningRawContentEvent,
-        AgentReasoningSectionBreakEvent, ApplyPatchApprovalRequestEvent, ErrorEvent, Event,
+        AgentReasoningSectionBreakEvent, ApplyPatchApprovalRequestEvent, ErrorEvent,
         ExecApprovalRequestEvent, ExecCommandBeginEvent, FileChange, McpInvocation,
         McpToolCallBeginEvent, McpToolCallEndEvent, PatchApplyBeginEvent, PatchApplyEndEvent,
         ReviewDecision, StreamErrorEvent, TaskCompleteEvent, TaskStartedEvent, TurnAbortedEvent,
         UserMessageEvent, WebSearchBeginEvent, WebSearchEndEvent,
     },
 };
-use codex_protocol::{
-    ConversationId,
-    config_types::ReasoningEffort,
-    protocol::{EventMsg, InputItem, Op},
-};
+use codex_protocol::{ConversationId, protocol::EventMsg};
 use itertools::Itertools;
 use mcp_types::CallToolResult;
 use std::{
@@ -45,12 +37,13 @@ use std::{
     collections::HashMap,
     path::PathBuf,
     rc::Rc,
-    sync::{Arc, LazyLock, OnceLock},
+    sync::{Arc, OnceLock},
 };
-use tokio::sync::{mpsc, oneshot};
+use tokio::sync::mpsc;
 use tracing::{debug, error, info};
 
-static APPROVAL_PRESETS: LazyLock<Vec<ApprovalPreset>> = LazyLock::new(builtin_approval_presets);
+use crate::conversation::ConversationHandle;
+
 pub static ACP_CLIENT: OnceLock<AgentSideConnection> = OnceLock::new();
 
 /// The Codex implementation of the ACP Agent trait.
@@ -634,81 +627,6 @@ impl CodexAgent {
     }
 }
 
-fn build_prompt_items(prompt: Vec<ContentBlock>) -> Vec<InputItem> {
-    prompt
-        .into_iter()
-        .filter_map(|block| match block {
-            ContentBlock::Text(text_block) => Some(InputItem::Text {
-                text: text_block.text.clone(),
-            }),
-            ContentBlock::Image(image_block) => {
-                // Convert to data URI if needed
-                if let Some(uri) = &image_block.uri {
-                    Some(InputItem::Image {
-                        image_url: uri.clone(),
-                    })
-                } else {
-                    // Base64 data
-                    let data_uri = format!(
-                        "data:{};base64,{}",
-                        image_block.mime_type.clone(),
-                        image_block.data.clone()
-                    );
-                    Some(InputItem::Image {
-                        image_url: data_uri,
-                    })
-                }
-            }
-            ContentBlock::ResourceLink(ResourceLink {
-                annotations: _,
-                description: _,
-                mime_type: _,
-                name,
-                size: _,
-                title: _,
-                uri,
-                meta: _,
-            }) => Some(InputItem::Text {
-                text: format_uri_as_link(Some(name), uri),
-            }),
-            ContentBlock::Resource(EmbeddedResource {
-                annotations: _,
-                resource:
-                    EmbeddedResourceResource::TextResourceContents(TextResourceContents {
-                        mime_type: _,
-                        text,
-                        uri,
-                        meta: _,
-                    }),
-                meta: _,
-            }) => Some(InputItem::Text {
-                text: format!(
-                    "{}\n<context ref=\"{uri}\">\n${text}\n</context>",
-                    format_uri_as_link(None, uri.clone())
-                ),
-            }),
-            // Skip other content types for now
-            ContentBlock::Audio(..) | ContentBlock::Resource(..) => None,
-        })
-        .collect()
-}
-
-fn format_uri_as_link(name: Option<String>, uri: String) -> String {
-    if let Some(name) = name
-        && !name.is_empty()
-    {
-        format!("[@{name}]({uri})")
-    } else if let Some(path) = uri.strip_prefix("file://") {
-        let name = path.split('/').next_back().unwrap_or(path);
-        format!("[@{name}]({uri})")
-    } else if uri.starts_with("zed://") {
-        let name = uri.split('/').next_back().unwrap_or(&uri);
-        format!("[@{name}]({uri})")
-    } else {
-        uri
-    }
-}
-
 #[async_trait::async_trait(?Send)]
 impl Agent for CodexAgent {
     async fn initialize(&self, request: InitializeRequest) -> Result<InitializeResponse, Error> {
@@ -1258,490 +1176,6 @@ fn convert_annotations(annotations: mcp_types::Annotations) -> Annotations {
         last_modified: annotations.last_modified,
         priority: annotations.priority,
         meta: None,
-    }
-}
-
-enum ConversationMessage {
-    Load {
-        response_tx: oneshot::Sender<Result<LoadSessionResponse, Error>>,
-    },
-    Prompt {
-        request: PromptRequest,
-        response_tx: oneshot::Sender<Result<(String, mpsc::UnboundedReceiver<EventMsg>), Error>>,
-    },
-    ExecApproval {
-        submission_id: String,
-        decision: ReviewDecision,
-        response_tx: oneshot::Sender<Result<(), Error>>,
-    },
-    PatchApproval {
-        submission_id: String,
-        decision: ReviewDecision,
-        response_tx: oneshot::Sender<Result<(), Error>>,
-    },
-    SetMode {
-        mode: SessionModeId,
-        response_tx: oneshot::Sender<Result<(), Error>>,
-    },
-    SetModel {
-        model: ModelId,
-        response_tx: oneshot::Sender<Result<(), Error>>,
-    },
-    Cancel {
-        response_tx: oneshot::Sender<Result<(), Error>>,
-    },
-}
-
-struct ConversationHandle {
-    /// A sender for interacting with the conversation.
-    message_tx: mpsc::UnboundedSender<ConversationMessage>,
-    /// A handle to the spawned task.
-    _handle: tokio::task::JoinHandle<()>,
-}
-
-impl ConversationHandle {
-    fn new(
-        conversation: Arc<CodexConversation>,
-        config: Config,
-        model_presets: Arc<Vec<ModelPreset>>,
-    ) -> Self {
-        let (message_tx, message_rx) = mpsc::unbounded_channel();
-
-        let actor = ConversationActor::new(conversation.clone(), config, model_presets, message_rx);
-        let handle = tokio::spawn(actor.spawn());
-
-        Self {
-            message_tx,
-            _handle: handle,
-        }
-    }
-
-    async fn load(&self) -> Result<LoadSessionResponse, Error> {
-        let (response_tx, response_rx) = oneshot::channel();
-
-        let message = ConversationMessage::Load { response_tx };
-        drop(self.message_tx.send(message));
-
-        response_rx
-            .await
-            .map_err(|e| Error::internal_error().with_data(e.to_string()))?
-    }
-
-    async fn prompt(
-        &self,
-        request: PromptRequest,
-    ) -> Result<(String, mpsc::UnboundedReceiver<EventMsg>), Error> {
-        let (response_tx, response_rx) = oneshot::channel();
-
-        let message = ConversationMessage::Prompt {
-            request,
-            response_tx,
-        };
-        drop(self.message_tx.send(message));
-
-        response_rx
-            .await
-            .map_err(|e| Error::internal_error().with_data(e.to_string()))?
-    }
-
-    async fn exec_approval(
-        &self,
-        submission_id: String,
-        decision: ReviewDecision,
-    ) -> Result<(), Error> {
-        let (response_tx, response_rx) = oneshot::channel();
-
-        let message = ConversationMessage::ExecApproval {
-            submission_id,
-            decision,
-            response_tx,
-        };
-        drop(self.message_tx.send(message));
-
-        response_rx
-            .await
-            .map_err(|e| Error::internal_error().with_data(e.to_string()))?
-    }
-
-    async fn patch_approval(
-        &self,
-        submission_id: String,
-        decision: ReviewDecision,
-    ) -> Result<(), Error> {
-        let (response_tx, response_rx) = oneshot::channel();
-
-        let message = ConversationMessage::PatchApproval {
-            submission_id,
-            decision,
-            response_tx,
-        };
-        drop(self.message_tx.send(message));
-
-        response_rx
-            .await
-            .map_err(|e| Error::internal_error().with_data(e.to_string()))?
-    }
-
-    async fn set_mode(&self, mode: SessionModeId) -> Result<(), Error> {
-        let (response_tx, response_rx) = oneshot::channel();
-
-        let message = ConversationMessage::SetMode { mode, response_tx };
-        drop(self.message_tx.send(message));
-
-        response_rx
-            .await
-            .map_err(|e| Error::internal_error().with_data(e.to_string()))?
-    }
-
-    async fn set_model(&self, model: ModelId) -> Result<(), Error> {
-        let (response_tx, response_rx) = oneshot::channel();
-
-        let message = ConversationMessage::SetModel { model, response_tx };
-        drop(self.message_tx.send(message));
-
-        response_rx
-            .await
-            .map_err(|e| Error::internal_error().with_data(e.to_string()))?
-    }
-
-    async fn cancel(&self) -> Result<(), Error> {
-        let (response_tx, response_rx) = oneshot::channel();
-
-        let message = ConversationMessage::Cancel { response_tx };
-        drop(self.message_tx.send(message));
-
-        response_rx
-            .await
-            .map_err(|e| Error::internal_error().with_data(e.to_string()))?
-    }
-}
-
-enum Submission {
-    Prompt(PromptSubmission),
-}
-
-impl Submission {
-    fn is_active(&self) -> bool {
-        match self {
-            Submission::Prompt(submission) => !submission.sender.is_closed(),
-        }
-    }
-
-    fn handle_event(&mut self, event: EventMsg) {
-        match self {
-            Submission::Prompt(submission) => submission.handle_event(event),
-        }
-    }
-}
-
-struct PromptSubmission {
-    sender: mpsc::UnboundedSender<EventMsg>,
-}
-
-impl PromptSubmission {
-    fn new(sender: mpsc::UnboundedSender<EventMsg>) -> Self {
-        Self { sender }
-    }
-
-    fn handle_event(&mut self, event: EventMsg) {
-        self.sender.send(event).ok();
-    }
-}
-
-struct ConversationActor {
-    /// The conversation associated with this task.
-    conversation: Arc<CodexConversation>,
-    /// The configuration for the conversation.
-    config: Config,
-    /// The model presets for the conversation.
-    model_presets: Arc<Vec<ModelPreset>>,
-    /// A sender for each interested `Op` submission that needs events routed.
-    submissions: HashMap<String, Submission>,
-    /// A receiver for incoming conversation messages.
-    message_rx: mpsc::UnboundedReceiver<ConversationMessage>,
-}
-
-impl ConversationActor {
-    fn new(
-        conversation: Arc<CodexConversation>,
-        config: Config,
-        model_presets: Arc<Vec<ModelPreset>>,
-        message_rx: mpsc::UnboundedReceiver<ConversationMessage>,
-    ) -> Self {
-        Self {
-            conversation,
-            config,
-            model_presets,
-            submissions: HashMap::new(),
-            message_rx,
-        }
-    }
-
-    async fn spawn(mut self) {
-        loop {
-            tokio::select! {
-                biased;
-                message = self.message_rx.recv() => match message {
-                    Some(message) => self.handle_message(message).await,
-                    None => break,
-                },
-                event = self.conversation.next_event() => match event {
-                    Ok(event) => self.handle_event(event),
-                    Err(e) => {
-                        error!("Error getting next event: {:?}", e);
-                        break;
-                    }
-                }
-            }
-            // Litter collection of senders with no receivers
-            self.submissions.retain(|_, sender| sender.is_active());
-        }
-    }
-
-    async fn handle_message(&mut self, message: ConversationMessage) {
-        match message {
-            ConversationMessage::Load { response_tx } => {
-                let result = self.handle_load().await;
-                drop(response_tx.send(result));
-            }
-            ConversationMessage::Prompt {
-                request,
-                response_tx,
-            } => {
-                let result = self.handle_prompt(request).await;
-                drop(response_tx.send(result));
-            }
-            ConversationMessage::ExecApproval {
-                submission_id,
-                decision,
-                response_tx,
-            } => {
-                let result = self.handle_exec_approval(submission_id, decision).await;
-                drop(response_tx.send(result));
-            }
-            ConversationMessage::PatchApproval {
-                submission_id,
-                decision,
-                response_tx,
-            } => {
-                let result = self.handle_patch_approval(submission_id, decision).await;
-                drop(response_tx.send(result));
-            }
-            ConversationMessage::SetMode { mode, response_tx } => {
-                let result = self.handle_set_mode(mode).await;
-                drop(response_tx.send(result));
-            }
-            ConversationMessage::SetModel { model, response_tx } => {
-                let result = self.handle_set_model(model).await;
-                drop(response_tx.send(result));
-            }
-            ConversationMessage::Cancel { response_tx } => {
-                let result = self.handle_cancel().await;
-                drop(response_tx.send(result));
-            }
-        }
-    }
-
-    fn modes(&self) -> Option<SessionModeState> {
-        let current_mode_id = APPROVAL_PRESETS
-            .iter()
-            .find(|preset| {
-                preset.approval == self.config.approval_policy
-                    && preset.sandbox == self.config.sandbox_policy
-            })
-            .map(|preset| SessionModeId(preset.id.into()))?;
-
-        Some(SessionModeState {
-            current_mode_id,
-            available_modes: APPROVAL_PRESETS
-                .iter()
-                .map(|preset| SessionMode {
-                    id: SessionModeId(preset.id.into()),
-                    name: preset.label.to_owned(),
-                    description: Some(preset.description.to_owned()),
-                    meta: None,
-                })
-                .collect(),
-            meta: None,
-        })
-    }
-
-    fn find_model_preset(&self) -> Option<&ModelPreset> {
-        if let Some(preset) = self.model_presets.iter().find(|preset| {
-            preset.model == self.config.model && preset.effort == self.config.model_reasoning_effort
-        }) {
-            return Some(preset);
-        }
-
-        // If we didn't find it, and it is set to none, see if we can find one with the default value
-        if self.config.model_reasoning_effort.is_none()
-            && let Some(preset) = self.model_presets.iter().find(|preset| {
-                preset.model == self.config.model
-                    && preset.effort == Some(ReasoningEffort::default())
-            })
-        {
-            return Some(preset);
-        }
-
-        None
-    }
-
-    fn models(&self) -> Result<SessionModelState, Error> {
-        let current_model_id = self
-            .find_model_preset()
-            .map(|preset| ModelId(preset.id.into()))
-            .ok_or_else(|| {
-                anyhow::anyhow!("No valid model preset for model {}", self.config.model)
-            })?;
-
-        let available_models = self
-            .model_presets
-            .iter()
-            .map(|preset| ModelInfo {
-                model_id: ModelId(preset.id.into()),
-                name: preset.label.into(),
-                description: Some(
-                    preset
-                        .description
-                        .strip_prefix("— ")
-                        .unwrap_or(preset.description)
-                        .into(),
-                ),
-                meta: None,
-            })
-            .collect();
-
-        Ok(SessionModelState {
-            current_model_id,
-            available_models,
-            meta: None,
-        })
-    }
-
-    async fn handle_load(&mut self) -> Result<LoadSessionResponse, Error> {
-        Ok(LoadSessionResponse {
-            modes: self.modes(),
-            models: Some(self.models()?),
-            meta: None,
-        })
-    }
-
-    async fn handle_prompt(
-        &mut self,
-        request: PromptRequest,
-    ) -> Result<(String, mpsc::UnboundedReceiver<EventMsg>), Error> {
-        let (session_tx, session_rx) = mpsc::unbounded_channel();
-        let items = build_prompt_items(request.prompt);
-
-        let submission_id = match self.conversation.submit(Op::UserInput { items }).await {
-            Ok(submission_id) => submission_id,
-            Err(e) => {
-                error!("Failed to submit prompt: {e:?}");
-                return Err(Error::internal_error());
-            }
-        };
-
-        self.submissions.insert(
-            submission_id.clone(),
-            Submission::Prompt(PromptSubmission::new(session_tx)),
-        );
-
-        Ok((submission_id, session_rx))
-    }
-
-    async fn handle_exec_approval(
-        &mut self,
-        submission_id: String,
-        decision: ReviewDecision,
-    ) -> Result<(), Error> {
-        self.conversation
-            .submit(Op::ExecApproval {
-                id: submission_id,
-                decision,
-            })
-            .await
-            .map_err(|e| Error::from(anyhow::anyhow!(e)))?;
-        Ok(())
-    }
-
-    async fn handle_patch_approval(
-        &mut self,
-        submission_id: String,
-        decision: ReviewDecision,
-    ) -> Result<(), Error> {
-        self.conversation
-            .submit(Op::PatchApproval {
-                id: submission_id,
-                decision,
-            })
-            .await
-            .map_err(|e| Error::from(anyhow::anyhow!(e)))?;
-        Ok(())
-    }
-
-    async fn handle_set_mode(&mut self, mode: SessionModeId) -> Result<(), Error> {
-        let preset = APPROVAL_PRESETS
-            .iter()
-            .find(|preset| mode.0.as_ref() == preset.id)
-            .ok_or_else(Error::invalid_params)?;
-
-        self.conversation
-            .submit(Op::OverrideTurnContext {
-                cwd: None,
-                approval_policy: Some(preset.approval),
-                sandbox_policy: Some(preset.sandbox.clone()),
-                model: None,
-                effort: None,
-                summary: None,
-            })
-            .await
-            .map_err(|e| Error::from(anyhow::anyhow!(e)))?;
-
-        self.config.approval_policy = preset.approval;
-        self.config.sandbox_policy = preset.sandbox.clone();
-
-        Ok(())
-    }
-
-    async fn handle_set_model(&mut self, model: ModelId) -> Result<(), Error> {
-        let preset = self
-            .model_presets
-            .iter()
-            .find(|p| p.id == model.0.as_ref())
-            .ok_or_else(|| Error::invalid_params().with_data("Model not found"))?;
-
-        self.conversation
-            .submit(Op::OverrideTurnContext {
-                cwd: None,
-                approval_policy: None,
-                sandbox_policy: None,
-                model: Some(preset.model.into()),
-                effort: Some(preset.effort),
-                summary: None,
-            })
-            .await
-            .map_err(|e| Error::from(anyhow::anyhow!(e)))?;
-
-        self.config.model = preset.model.into();
-        self.config.model_reasoning_effort = preset.effort;
-
-        Ok(())
-    }
-
-    async fn handle_cancel(&mut self) -> Result<(), Error> {
-        self.conversation
-            .submit(Op::Interrupt)
-            .await
-            .map_err(|e| Error::from(anyhow::anyhow!(e)))?;
-        Ok(())
-    }
-
-    fn handle_event(&mut self, Event { id, msg }: Event) {
-        if let Some(submission) = self.submissions.get_mut(&id) {
-            submission.handle_event(msg);
-        } else {
-            error!("Received event for unknown submission ID: {}", id);
-        }
     }
 }
